Add Tests (#1)

+++ deepgrp/optimization.py
@@ -1,20 +1,24 @@
 """deepgrp.optimization modul for optimisation with hyperopt"""
 
-from typing import Any, Dict, Callable, Union
+import logging
 import pickle
-from os import path, PathLike
 import shutil
+from os import PathLike, path
+from typing import Any, Callable, Dict, Union
+
 import numpy as np
-from hyperopt import fmin, tpe, Trials, STATUS_OK, STATUS_FAIL
-from tensorboard.plugins.hparams import api as hp
 import tensorflow as tf
+from hyperopt import STATUS_FAIL, STATUS_OK, Trials, fmin, tpe
+from tensorboard.plugins.hparams import api as hp
 from tensorflow import summary as tfsum
 from tensorflow.keras import backend as K  # pylint: disable=import-error
+
+from deepgrp.model import Options, create_logdir, create_model
+import deepgrp.prediction as dgpred
 from deepgrp.preprocessing import Data
-from deepgrp.model import create_model, create_logdir, Options
 from deepgrp.training import training
-from deepgrp.prediction import predict_complete, filter_segments
-from deepgrp.prediction import calculate_metrics
+
+_LOGGER = logging.getLogger(__name__)
 
 
 def _update_options(options: Options, dictionary: Dict[str, Any]) -> Options:
@@ -25,10 +29,9 @@ def _update_options(options: Options, dictionary: Dict[str, Any]) -> Options:
     return options
 
 
-def build_and_optimize(train_data: Data, val_data: Data, step_size: int,
-                       options: Options,
-                       options_dict: Dict[str, Union[str, float]]
-                       ) -> Dict[str, Any]:
+def build_and_optimize(
+        train_data: Data, val_data: Data, step_size: int, options: Options,
+        options_dict: Dict[str, Union[str, float]]) -> Dict[str, Any]:
     """Builds an DeepGRP model with updated options,
     trains it and validates it. Used for hyperopt optimization.
 
@@ -47,21 +50,21 @@ def build_and_optimize(train_data: Data, val_data: Data, step_size: int,
     options = _update_options(options, options_dict)
     logdir = create_logdir(options)
 
-    def _train_test(model):
+    def _train_test(model):  # pragma: no cover
         extra_callback = [hp.KerasCallback(logdir, options_dict)]
         training((train_data, val_data), options, model, logdir,
                  extra_callback)
         K.clear_session()
-        predictions = predict_complete(step_size,
-                                       options,
-                                       logdir,
-                                       val_data,
-                                       use_mss=True)
+        predictions = dgpred.predict_complete(step_size,
+                                              options,
+                                              logdir,
+                                              val_data,
+                                              use_mss=True)
         K.clear_session()
         is_not_na = np.logical_not(np.isnan(predictions[:, 0]))
         predictions_class = predictions[is_not_na].argmax(axis=1)
-        filter_segments(predictions_class, options.min_mss_len)
-        _, metrics = calculate_metrics(
+        dgpred.filter_segments(predictions_class, options.min_mss_len)
+        _, metrics = dgpred.calculate_metrics(
             predictions_class, val_data.truelbl[:, is_not_na].argmax(axis=0))
         return metrics
 
@@ -83,7 +86,11 @@ def build_and_optimize(train_data: Data, val_data: Data, step_size: int,
                      metrics['MCC'],
                      step=0,
                      description="Matthews correlation coefficient")
-        file_writer.close()
+    except Exception as err:  # pylint: disable=broad-except
+        _LOGGER.exception("Error occurred while training")
+        results["error"] = str(err)
+        results["status"] = STATUS_FAIL
+    else:
         results["logdir"] = logdir
         results["loss"] = -1 * metrics['MCC']
 
@@ -92,17 +99,15 @@ def build_and_optimize(train_data: Data, val_data: Data, step_size: int,
         if np.isnan(results["loss"]):
             results["status"] = STATUS_FAIL
             results["loss"] = np.inf
-    except Exception as err:  # pylint: disable=broad-except
-        print(err)
-        results["error"] = str(err)
-        results["status"] = STATUS_FAIL
-        if results["logdir"]:
-            shutil.rmtree(results["logdir"])
+    finally:
+        file_writer.close()
+    if results["status"] == STATUS_FAIL and results["logdir"]:
+        shutil.rmtree(results["logdir"])
     return results
 
 
-def run_a_trial(space: Dict[str, Any],
-                objective: Callable[[Dict[str, Any]], Dict[str, Any]],
+def run_a_trial(space: Dict[str, Any], objective: Callable[[Dict[str, Any]],
+                                                           Dict[str, Any]],
                 project_root_dir: PathLike, max_evals: int) -> int:
     """Runs a hyperopt TPE meta optimisation step. Restores previous run,
      if `results.pkl` is available.
@@ -122,19 +127,20 @@ def run_a_trial(space: Dict[str, Any],
     """
     nb_evals = max_evals
 
-    print("Attempt to resume a past training if it exists:")
+    _LOGGER.info("Attempt to resume a past training if it exists:")
     results_path = path.join(project_root_dir, "results.pkl")
 
     try:
         with open(results_path, "rb") as file:
             trials = pickle.load(file)
-        print("Found saved Trials! Loading...")
-        max_evals = len(trials.trials) + nb_evals
-        print("Rerunning from {} trials to add another one.".format(
-            len(trials.trials)))
     except FileNotFoundError:
         trials = Trials()
-        print("Starting from scratch: new trials.")
+        _LOGGER.info("Starting from scratch: new trials.")
+    else:
+        _LOGGER.warning("Found saved Trials! Loading...")
+        max_evals = len(trials.trials) + nb_evals
+        _LOGGER.info("Rerunning from %d trials to add another one.",
+                     len(trials.trials))
 
     fmin(objective,
          space,
